requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
gitpython
requests
watchfiles